feat(options): move code block appearance settings to "Text notes"
